Extract out sync self-ping
